backend/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
certifi==2020.12.5
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.7.0
django-heroku==0.3.1
django-tastypie==0.14.3
djangorestframework==3.12.4
filelock==3.0.12
gunicorn==20.1.0
pipenv==2020.11.15
psycopg2==2.8.6
python-dateutil==2.8.1
python-decouple==3.4
python-mimeparse==1.6.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
virtualenv==20.4.4
virtualenv-clone==0.5.4
whitenoise==5.2.0
